+++ src/SCEMonteCarlo.jl
@@ -72,6 +72,7 @@ export ReducedCell, reduce_cell
 public resolve_kt
 public ScaledTerm, SpinConfig, site_index, site_atom
 public site_coeffs!, delta_energy, site_gradient, energy_gradient, energy_gradient!
+public philox_block, philox_normal2
 public LogBinner, BinStore, jackknife, std_error, tau_int, bin_means
 public ChainState, SweepScratch, metropolis_sweep!, overrelaxation_sweep!
 public to_matrix, from_matrix